fix: move AppSettings.cs to `settings` folder
